Networking and API Challenge(Start)
What I will make ?
The objective of this challenge is to apply the skills you learned in the Clima tutorial and make a Bitcoin price reporter. We are going to revisit how to use URLSession to make HTTP requests and JSONDecoder for JSON parsing. This will be the most advanced challenge you will face to date. We believe in you!